changes for third review
